+++ Source/Core/Visualization/Renderer/PointRendererGLSL.cpp
@@ -49,19 +49,39 @@ namespace kvs
 namespace glsl
 {
 
+void PointRenderer::BufferObject::set( const kvs::PointObject* point )
+{
+    const bool has_normal = point->normals().size() > 0;
+    auto coords = point->coords();
+    auto colors = ::VertexColors( point );
+    auto normals = point->normals();
+    m_manager.setVertexArray( coords, 3 );
+    m_manager.setColorArray( colors, 3 );
+    if ( has_normal ) { m_manager.setNormalArray( normals ); }
+}
+
+void PointRenderer::BufferObject::draw( const kvs::PointObject* point )
+{
+    kvs::VertexBufferObjectManager::Binder bind( m_manager );
+    {
+        const size_t nvertices = point->numberOfVertices();
+        m_manager.drawArrays( GL_POINTS, 0, nvertices );
+    }
+}
+
 /*===========================================================================*/
 /**
  *  @brief  Constructs a new PointRenderer class.
  */
 /*===========================================================================*/
 PointRenderer::PointRenderer():
+    m_vert_file( "shader.vert" ),
+    m_frag_file( "shader.frag" ),
     m_width( 0 ),
     m_height( 0 ),
     m_object( NULL ),
-    m_has_normal( false ),
-    m_shader( NULL )
+    m_shading_model( new kvs::Shader::Lambert() )
 {
-    this->setShader( kvs::Shader::Lambert() );
 }
 
 /*===========================================================================*/
@@ -71,7 +91,7 @@ PointRenderer::PointRenderer():
 /*===========================================================================*/
 PointRenderer::~PointRenderer()
 {
-    if ( m_shader ) { delete m_shader; }
+    if ( m_shading_model ) { delete m_shading_model; }
 }
 
 /*===========================================================================*/
@@ -84,60 +104,35 @@ PointRenderer::~PointRenderer()
 /*===========================================================================*/
 void PointRenderer::exec( kvs::ObjectBase* object, kvs::Camera* camera, kvs::Light* light )
 {
-    kvs::PointObject* point = kvs::PointObject::DownCast( object );
-    m_has_normal = point->numberOfNormals() > 0;
-    if ( !m_has_normal ) setEnabledShading( false );
-
     BaseClass::startTimer();
     kvs::OpenGL::WithPushedAttrib p( GL_ALL_ATTRIB_BITS );
-    kvs::OpenGL::Enable( GL_DEPTH_TEST );
-    kvs::OpenGL::Enable( GL_POINT_SMOOTH ); // Rounded shape.
 
     const size_t width = camera->windowWidth();
     const size_t height = camera->windowHeight();
-    const bool window_created = m_width == 0 && m_height == 0;
-    if ( window_created )
+
+    if ( this->isWindowCreated() )
     {
-        m_width = width;
-        m_height = height;
-        m_object = object;
-        this->create_shader_program();
-        this->create_buffer_object( point );
+        this->setWindowSize( width, height );
+        this->createBufferObject( object );
+        this->createShaderProgram();
     }
 
-    const bool window_resized = m_width != width || m_height != height;
-    if ( window_resized )
+    if ( this->isWindowResized( width, height ) )
     {
-        m_width = width;
-        m_height = height;
+        this->setWindowSize( width, height );
     }
 
-    const bool object_changed = m_object != object;
-    if ( object_changed )
+    if ( this->isObjectChanged( object ) )
     {
-        m_object = object;
-        m_shader_program.release();
-        m_vbo_manager.release();
-        this->create_shader_program();
-        this->create_buffer_object( point );
+        this->updateBufferObject( object );
+        this->updateShaderProgram();
     }
 
-    kvs::VertexBufferObjectManager::Binder bind1( m_vbo_manager );
-    kvs::ProgramObject::Binder bind2( m_shader_program );
-    {
-        const kvs::Mat4 M = kvs::OpenGL::ModelViewMatrix();
-        const kvs::Mat4 PM = kvs::OpenGL::ProjectionMatrix() * M;
-        const kvs::Mat3 N = kvs::Mat3( M[0].xyz(), M[1].xyz(), M[2].xyz() );
-        m_shader_program.setUniform( "ModelViewMatrix", M );
-        m_shader_program.setUniform( "ModelViewProjectionMatrix", PM );
-        m_shader_program.setUniform( "NormalMatrix", N );
+    this->setupShaderProgram();
 
-        const float dpr = camera->devicePixelRatio();
-        kvs::OpenGL::SetPointSize( point->size() * dpr );
-
-        const size_t nvertices = point->numberOfVertices();
-        m_vbo_manager.drawArrays( GL_POINTS, 0, nvertices );
-    }
+    m_shader_program.bind();
+    this->drawBufferObject( camera );
+    m_shader_program.unbind();
 
     BaseClass::stopTimer();
 }
@@ -147,13 +142,13 @@ void PointRenderer::exec( kvs::ObjectBase* object, kvs::Camera* camera, kvs::Lig
  *  @brief  Creates shader program.
  */
 /*===========================================================================*/
-void PointRenderer::create_shader_program()
+void PointRenderer::createShaderProgram()
 {
-    kvs::ShaderSource vert( "shader.vert" );
-    kvs::ShaderSource frag( "shader.frag" );
+    kvs::ShaderSource vert( this->vertexShaderFile() );
+    kvs::ShaderSource frag( this->fragmentShaderFile() );
     if ( isEnabledShading() )
     {
-        switch ( m_shader->type() )
+        switch ( m_shading_model->type() )
         {
         case kvs::Shader::LambertShading: frag.define("ENABLE_LAMBERT_SHADING"); break;
         case kvs::Shader::PhongShading: frag.define("ENABLE_PHONG_SHADING"); break;
@@ -168,13 +163,29 @@ void PointRenderer::create_shader_program()
     }
 
     m_shader_program.build( vert, frag );
-    m_shader_program.bind();
-    m_shader_program.setUniform( "shading.Ka", m_shader->Ka );
-    m_shader_program.setUniform( "shading.Kd", m_shader->Kd );
-    m_shader_program.setUniform( "shading.Ks", m_shader->Ks );
-    m_shader_program.setUniform( "shading.S",  m_shader->S );
+}
+
+void PointRenderer::updateShaderProgram()
+{
+    m_shader_program.release();
+    this->createShaderProgram();
+}
+
+void PointRenderer::setupShaderProgram()
+{
+    kvs::ProgramObject::Binder bind( m_shader_program );
+    m_shader_program.setUniform( "shading.Ka", m_shading_model->Ka );
+    m_shader_program.setUniform( "shading.Kd", m_shading_model->Kd );
+    m_shader_program.setUniform( "shading.Ks", m_shading_model->Ks );
+    m_shader_program.setUniform( "shading.S",  m_shading_model->S );
     m_shader_program.setUniform( "offset", 0 );
-    m_shader_program.unbind();
+
+    const kvs::Mat4 M = kvs::OpenGL::ModelViewMatrix();
+    const kvs::Mat4 PM = kvs::OpenGL::ProjectionMatrix() * M;
+    const kvs::Mat3 N = kvs::Mat3( M[0].xyz(), M[1].xyz(), M[2].xyz() );
+    m_shader_program.setUniform( "ModelViewMatrix", M );
+    m_shader_program.setUniform( "ModelViewProjectionMatrix", PM );
+    m_shader_program.setUniform( "NormalMatrix", N );
 }
 
 /*===========================================================================*/
@@ -183,16 +194,29 @@ void PointRenderer::create_shader_program()
  *  @param  point [in] pointer to the point object
  */
 /*===========================================================================*/
-void PointRenderer::create_buffer_object( const kvs::PointObject* point )
+void PointRenderer::createBufferObject( const kvs::ObjectBase* object )
 {
-    kvs::ValueArray<kvs::Real32> coords = point->coords();
-    kvs::ValueArray<kvs::UInt8> colors = ::VertexColors( point );
-    kvs::ValueArray<kvs::Real32> normals = point->normals();
+    m_object = object;
+    m_buffer_object.set( kvs::PointObject::DownCast( object ) );
+    m_buffer_object.create();
+}
 
-    m_vbo_manager.setVertexArray( coords, 3 );
-    m_vbo_manager.setColorArray( colors, 3 );
-    if ( normals.size() > 0 ) { m_vbo_manager.setNormalArray( normals ); }
-    m_vbo_manager.create();
+void PointRenderer::updateBufferObject( const kvs::ObjectBase* object )
+{
+    m_buffer_object.release();
+    this->createBufferObject( object );
+}
+
+void PointRenderer::drawBufferObject( const kvs::Camera* camera )
+{
+    kvs::OpenGL::Enable( GL_DEPTH_TEST );
+    kvs::OpenGL::Enable( GL_POINT_SMOOTH ); // Rounded shape.
+
+    const auto* point = kvs::PointObject::DownCast( m_object );
+    const float dpr = camera->devicePixelRatio();
+    kvs::OpenGL::SetPointSize( point->size() * dpr );
+
+    m_buffer_object.draw( point );
 }
 
 } // end of namespace glsl