+++ Source/Core/Visualization/Renderer/StochasticLineRenderer.cpp
@@ -31,30 +31,6 @@ int RandomNumber()
     return C * R.randInteger();
 }
 
-/*===========================================================================*/
-/**
- *  @brief  Returns vertex-color array.
- *  @param  line [in] pointer to the line object
- */
-/*===========================================================================*/
-kvs::ValueArray<kvs::UInt8> VertexColors( const kvs::LineObject* line )
-{
-    if ( line->colorType() == kvs::LineObject::VertexColor ) return line->colors();
-
-    const size_t nvertices = line->numberOfVertices();
-    const kvs::RGBColor color = line->color();
-
-    kvs::ValueArray<kvs::UInt8> colors( nvertices * 3 );
-    for ( size_t i = 0; i < nvertices; i++ )
-    {
-        colors[ 3 * i + 0 ] = color.r();
-        colors[ 3 * i + 1 ] = color.g();
-        colors[ 3 * i + 2 ] = color.b();
-    }
-
-    return colors;
-}
-
 }
 
 namespace kvs
@@ -99,7 +75,6 @@ void StochasticLineRenderer::setOpacity( const kvs::UInt8 opacity )
 /*===========================================================================*/
 StochasticLineRenderer::Engine::Engine():
     m_line_opacity( 255 ),
-    m_has_connection( false ),
     m_line_offset( 0.0f )
 {
 }
@@ -112,7 +87,7 @@ StochasticLineRenderer::Engine::Engine():
 void StochasticLineRenderer::Engine::release()
 {
     m_shader_program.release();
-    m_vbo_manager.release();
+    m_buffer_object.release();
 }
 
 /*===========================================================================*/
@@ -128,11 +103,7 @@ void StochasticLineRenderer::Engine::create(
     kvs::Camera* camera,
     kvs::Light* light )
 {
-    kvs::LineObject* line = kvs::LineObject::DownCast( object );
-    kvs::LineObject::LineType type = line->lineType();
-    m_has_connection = line->numberOfConnections() > 0 &&
-        ( type == kvs::LineObject::Segment || type == kvs::LineObject::Uniline );
-
+    auto* line = kvs::LineObject::DownCast( object );
     attachObject( object );
     createRandomTexture();
     this->create_shader_program();
@@ -169,13 +140,12 @@ void StochasticLineRenderer::Engine::setup(
 {
     const kvs::Mat4 M = kvs::OpenGL::ModelViewMatrix();
     const kvs::Mat4 PM = kvs::OpenGL::ProjectionMatrix() * M;
-    m_shader_program.bind();
+    kvs::ProgramObject::Binder bind2( m_shader_program );
     m_shader_program.setUniform( "ModelViewProjectionMatrix", PM );
     m_shader_program.setUniform( "random_texture_size_inv", 1.0f / randomTextureSize() );
     m_shader_program.setUniform( "random_texture", 0 );
     m_shader_program.setUniform( "opacity", m_line_opacity / 255.0f );
     m_shader_program.setUniform( "line_offset", m_line_offset );
-    m_shader_program.unbind();
 }
 
 /*===========================================================================*/
@@ -191,52 +161,17 @@ void StochasticLineRenderer::Engine::draw(
     kvs::Camera* camera,
     kvs::Light* light )
 {
-    kvs::LineObject* line = kvs::LineObject::DownCast( object );
-
-    kvs::VertexBufferObjectManager::Binder bind1( m_vbo_manager );
+    const size_t size = randomTextureSize();
+    const int count = repetitionCount() * ::RandomNumber();
+    const float offset_x = static_cast<float>( ( count ) % size );
+    const float offset_y = static_cast<float>( ( count / size ) % size );
+    const kvs::Vec2 random_offset( offset_x, offset_y );
     kvs::ProgramObject::Binder bind2( m_shader_program );
+    m_shader_program.setUniform( "random_offset", random_offset );
+
     kvs::Texture::Binder bind3( randomTexture() );
-    {
-        kvs::OpenGL::WithEnabled d( GL_DEPTH_TEST );
-
-        const size_t size = randomTextureSize();
-        const int count = repetitionCount() * ::RandomNumber();
-        const float offset_x = static_cast<float>( ( count ) % size );
-        const float offset_y = static_cast<float>( ( count / size ) % size );
-        const kvs::Vec2 random_offset( offset_x, offset_y );
-        m_shader_program.setUniform( "random_offset", random_offset );
-
-        kvs::OpenGL::SetLineWidth( line->size() );
-        if ( m_has_connection )
-        {
-            const size_t nlines = line->numberOfConnections();
-            switch ( line->lineType() )
-            {
-            case kvs::LineObject::Uniline:
-                m_vbo_manager.drawElements( GL_LINE_STRIP, nlines );
-                break;
-            case kvs::LineObject::Segment:
-                m_vbo_manager.drawElements( GL_LINES, 2 * nlines );
-                break;
-            default:
-                break;
-            }
-        }
-        else
-        {
-            switch ( line->lineType() )
-            {
-            case kvs::LineObject::Polyline:
-                m_vbo_manager.drawArrays( GL_LINE_STRIP, m_first_array, m_count_array );
-                break;
-            case kvs::LineObject::Strip:
-                m_vbo_manager.drawArrays( GL_LINE_STRIP, 0, line->numberOfVertices() );
-                break;
-            default:
-                break;
-            }
-        }
-    }
+    auto* line = kvs::LineObject::DownCast( object );
+    m_buffer_object.draw( line );
 }
 
 /*===========================================================================*/
@@ -259,40 +194,20 @@ void StochasticLineRenderer::Engine::create_shader_program()
 /*===========================================================================*/
 void StochasticLineRenderer::Engine::create_buffer_object( const kvs::LineObject* line )
 {
-    if ( line->numberOfColors() != 1 && line->colorType() == kvs::LineObject::LineColor )
-    {
-        kvsMessageError() << "Not supported line color type." << std::endl;
-        return;
-    }
-
     const size_t nvertices = line->numberOfVertices();
+    const auto tex_size = randomTextureSize();
     kvs::ValueArray<kvs::UInt16> indices( nvertices * 2 );
     for ( size_t i = 0; i < nvertices; i++ )
     {
         const unsigned int count = i * 12347;
-        indices[ 2 * i + 0 ] = static_cast<kvs::UInt16>( ( count ) % randomTextureSize() );
-        indices[ 2 * i + 1 ] = static_cast<kvs::UInt16>( ( count / randomTextureSize() ) % randomTextureSize() );
+        indices[ 2 * i + 0 ] = static_cast<kvs::UInt16>( ( count ) % tex_size );
+        indices[ 2 * i + 1 ] = static_cast<kvs::UInt16>( ( count / tex_size ) % tex_size );
     }
-    kvs::ValueArray<kvs::Real32> coords = line->coords();
-    kvs::ValueArray<kvs::UInt8> colors = ::VertexColors( line );
 
-    m_vbo_manager.setVertexAttribArray( indices, m_shader_program.attributeLocation("random_index"), 2 );
-    m_vbo_manager.setVertexArray( coords, 3 );
-    m_vbo_manager.setColorArray( colors, 3 );
-    if ( m_has_connection ) { m_vbo_manager.setIndexArray( line->connections() ); }
-    m_vbo_manager.create();
-
-    if ( ( !m_has_connection ) && ( line->lineType() == kvs::LineObject::Polyline ) )
-    {
-        const kvs::UInt32* pconnections = line->connections().data();
-        m_first_array.allocate( line->numberOfConnections() );
-        m_count_array.allocate( m_first_array.size() );
-        for ( size_t i = 0; i < m_first_array.size(); ++i )
-        {
-            m_first_array[i] = pconnections[ 2 * i ];
-            m_count_array[i] = pconnections[ 2 * i + 1 ] - pconnections[ 2 * i ] + 1;
-        }
-    }
+    auto location = m_shader_program.attributeLocation( "random_index" );
+    m_buffer_object.manager().setVertexAttribArray( indices, location, 2 );
+    m_buffer_object.set( line );
+    m_buffer_object.create();
 }
 
 } // end of namespace kvs